+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # UNRELEASED
 
+-   Expose new constructor function `OlmMachine.openWithKey()`.
+    ([#119](https://github.com/matrix-org/matrix-rust-sdk-crypto-wasm/pull/119))
+
 -   Add `OlmMachine.importSecretsBundle()` and `OlmMachine.exportSecretsBundle()`
     methods as well as the `SecretsBundle` class to import end-to-end encryption
     secrets in a bundled manner.
@@ -11,7 +14,7 @@
     generate QR codes intended for the QR code login mechanism described in
     [MSC4108](https://github.com/matrix-org/matrix-spec-proposals/pull/4108).
 
--   Update matrix-rust-sdk to `35173347f`, which includes:
+-   Update matrix-rust-sdk to `7a887766c`, which includes:
 
     -   Add data types to parse the QR code data for the QR code login defined in
         [MSC4108](https://github.com/matrix-org/matrix-spec-proposals/pull/4108)